requirements.txt
openstacksdk>=0.9.18
hgijson>=1.4.3
python-novaclient>=9.1.0
